requirements.txt
streamlit
supabase
requests
schedule
resend
python-dotenv
